v23.2.14.15
- Címke alapú kamerafelvétel igénylése
- Nagyobb ISZR ikonok használata
- Igénylések eltüntetése az admin felületről, mostantól az alap felületen lehet módosítani a státuszt
