+++ Controllers/RequestsController.cs
@@ -3,6 +3,7 @@ using ISZR.Components;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.Rendering;
 using System.Drawing.Drawing2D;
+using System.Linq;
 
 namespace ISZR.Controllers
 {
@@ -93,6 +94,54 @@ namespace ISZR.Controllers
 			if (request == null) return NotFound();
 
 			// Oldal megjelenítése a kért igényléssel
+			return View(request);
+		}
+
+		// POST: Administrator/Details/5
+		[HttpPost]
+		[ValidateAntiForgeryToken]
+		public async Task<IActionResult> Details(int? id, string? status)
+		{
+			if (id == null || status == null || _context.Requests == null) return NotFound();
+
+			var request = await _context.Requests
+				.Include(r => r.RequestAuthor)
+				.Include(r => r.RequestFor)
+				.Include(r => r.RequestAuthor.Class)
+				.Include(r => r.RequestAuthor.Position)
+				.Include(r => r.RequestFor.Class)
+				.Include(r => r.RequestFor.Position)
+				.FirstOrDefaultAsync(m => m.RequestId == id);
+
+			if (request == null) return NotFound();
+
+			try
+			{
+				request.Status = status;
+				if (status == "Folyamatban")
+				{
+					request.ResolveDate = new DateTime();
+				}
+				else
+				{
+					request.ResolveDate = DateTime.Now;
+				}
+
+				_context.Update(request);
+				await _context.SaveChangesAsync();
+			}
+			catch (DbUpdateConcurrencyException)
+			{
+				if (!RequestExists(request.RequestId))
+				{
+					return NotFound();
+				}
+				else
+				{
+					throw;
+				}
+			}
+
 			return View(request);
 		}
 
@@ -361,7 +410,7 @@ namespace ISZR.Controllers
 		// POST: Kamerafelvétel lementése címkék alapján
 		[HttpPost]
 		[ValidateAntiForgeryToken]
-		public async Task<IActionResult> RecordsByTags(string inputWhy, string inputTags, string[] selectedCameras, [Bind("RequestId,Type,Status,Description,RequestAuthorId,RequestForId")] Request request)
+		public async Task<IActionResult> RecordsByTags(DateTime inputDate, string inputWhy, string inputTags, string[] selectedCameras, [Bind("RequestId,Type,Status,Description,RequestAuthorId,RequestForId")] Request request)
 		{
 			// Megadott értékek ellenőrzése
 			if (ModelState.IsValid)
@@ -379,16 +428,69 @@ namespace ISZR.Controllers
 				// Alapértelmezett státusz
 				request.Status = "Folyamatban";
 
-				string cameras = "";
 				// Kamerák sorrendbe helyezése
-				foreach (string camera in selectedCameras)
-				{
-					cameras += $", {camera}";
-				}
+				string cameras = string.Join(", ", selectedCameras);
 
 				// Igénylés leírása
 				request.Description = $"Kérem engedélyezni a kamerarendszerben rögzített adatok külső adattárolón történő tárolását, illetve felhasználását megkeresés alapján Bűnügyi vagy Felügyeleti szerv részére.<br /><br />" +
-					$"<dl>\r\n<dt><i class=\"far fa-eye\"></i> Lementésének oka</dt>\r\n<dd>{inputWhy}</dd>\r\n<dt><i class=\"fas fa-tags\"></i> Címkék megnevezése</dt>\r\n<dd>{inputTags}</dd>\r\n<dt><i class=\"fas fa-video\"></i> Megcímkézett kamerák</dt>\r\n<dd>{cameras}</dd>\r\n</dl>";
+					$"<dl>\r\n<dt><i class=\"far fa-eye\"></i> Lementésének oka</dt>\r\n<dd>{inputWhy}</dd>\r\n<dt><i class=\"fas fa-calendar\"></i> Esemény dátuma</dt>\r\n<dd>{inputDate.ToString("yyyy.MM.dd")}</dd>\r\n<dt><i class=\"fas fa-tags\"></i> Címkék megnevezése</dt>\r\n<dd>{inputTags}</dd>\r\n<dt><i class=\"fas fa-video\"></i> Megcímkézett kamerák</dt>\r\n<dd>{cameras}</dd>\r\n</dl>";
+
+				// Igénylés hozzáadása a rendszerhez
+				_context.Add(request);
+				await _context.SaveChangesAsync();
+
+				// Igénylés megnyítása
+				return RedirectToAction(nameof(Details), new { @id = request.RequestId });
+			}
+
+			// Amennyiben nem jók az értékek az oldal újratöltése
+			ViewData["RequestForId"] = new SelectList(_context.Users.OrderBy(u => u.DisplayName), "UserId", "DisplayName");
+			return View();
+		}
+
+		// GET: Kamerafelvétel lementése címkék alapján
+		public async Task<IActionResult> RecordsByTime()
+		{
+			// Az ISZR-ben nem megtalálható személyek kizására
+			if (!await Account.IsUserExists(_context)) return Forbid();
+
+			// Az oldalt csak ügyintézők tekinthetik meg
+			if (!Account.IsUgyintezo()) return Forbid();
+
+			ViewData["Cameras"] = new MultiSelectList(_context.Cameras.OrderBy(p => p.Name), "Name", "Name");
+
+			// Az oldal megjelenítése
+			return View();
+		}
+
+		// POST: Kamerafelvétel lementése címkék alapján
+		[HttpPost]
+		[ValidateAntiForgeryToken]
+		public async Task<IActionResult> RecordsByTime(DateTime inputDate, string inputWhy, string inputTags, string[] selectedCameras, [Bind("RequestId,Type,Status,Description,RequestAuthorId,RequestForId")] Request request)
+		{
+			// Megadott értékek ellenőrzése
+			if (ModelState.IsValid)
+			{
+				// Igénylést létrehozó személy azonosítója
+				request.RequestAuthorId = await RequestAuthorId();
+				request.RequestForId = request.RequestAuthorId;
+
+				// Igénylés létrehozásának dátuma
+				request.CreationDate = DateTime.Now;
+
+				// Igénylés típusa
+				request.Type = "Kamerafelvétel lementése";
+
+				// Alapértelmezett státusz
+				request.Status = "Folyamatban";
+
+				// Kamerák sorrendbe helyezése
+				string cameras = string.Join(", ", selectedCameras);
+
+				// Igénylés leírása
+				request.Description = $"Kérem engedélyezni a kamerarendszerben rögzített adatok külső adattárolón történő tárolását, illetve felhasználását megkeresés alapján Bűnügyi vagy Felügyeleti szerv részére.<br /><br />" +
+					$"<dl>\r\n<dt><i class=\"far fa-eye\"></i> Lementésének oka</dt>\r\n<dd>{inputWhy}</dd>\r\n<dt><i class=\"fas fa-calendar\"></i> Esemény dátuma</dt>\r\n<dd>{inputDate.ToString("yyyy.MM.dd")}</dd>\r\n<dt><i class=\"fas fa-tags\"></i> Címkék megnevezése</dt>\r\n<dd>{inputTags}</dd>\r\n<dt><i class=\"fas fa-video\"></i> Megcímkézett kamerák</dt>\r\n<dd>{cameras}</dd>\r\n</dl>";
+
 				// Igénylés hozzáadása a rendszerhez
 				_context.Add(request);
 				await _context.SaveChangesAsync();
