--- Views/Administrator/Index.cshtml
@@ -1,129 +1,0 @@
-﻿@model IEnumerable<ISZR.Models.Request>
-
-@{
-    Layout = "~/Views/Shared/_AdminLayout.cshtml";
-    ViewData["Title"] = "Igénylések karbantartása";
-}
-
-<div class="container-fluid">
-    <div class="row">
-        <div class="col-lg-3 col-sm-6 col-12">
-            <div class="info-box mb-3">
-                <span class="info-box-icon bg-info elevation-1"><i class="fas fa-tasks"></i></span>
-                <div class="info-box-content">
-                    <span class="info-box-text">Igénylés összesen</span>
-                    <span class="info-box-number">@ViewBag.All</span>
-                </div>
-            </div>
-        </div>
-        <div class="col-lg-3 col-sm-6 col-12">
-            <div class="info-box mb-3">
-                <span class="info-box-icon bg-info elevation-1"><i class="fas fa-comment-dots"></i></span>
-                <div class="info-box-content">
-                    <span class="info-box-text">Folyamatban</span>
-                    <span class="info-box-number">@ViewBag.InProgress</span>
-                </div>
-            </div>
-        </div>
-        <div class="col-lg-3 col-sm-6 col-12">
-            <div class="info-box mb-3">
-                <span class="info-box-icon bg-success elevation-1"><i class="fas fa-check"></i></span>
-                <div class="info-box-content">
-                    <span class="info-box-text">Végrehajtva</span>
-                    <span class="info-box-number">@ViewBag.Done</span>
-                </div>
-            </div>
-        </div>
-        <div class="col-lg-3 col-sm-6 col-12">
-            <div class="info-box mb-3">
-                <span class="info-box-icon bg-danger elevation-1"><i class="fas fa-ban"></i></span>
-                <div class="info-box-content">
-                    <span class="info-box-text">Elutasítva</span>
-                    <span class="info-box-number">@ViewBag.Denied</span>
-                </div>
-            </div>
-        </div>
-    </div>
-    <div class="row">
-        <div class="col-12">
-            <div class="card">
-                <div class="card-body p-0">
-                    <table class="table table-striped">
-                        <thead class="bg-info">
-                            <tr>
-                                <th>Igénylési azonosító</th>
-                                <th>Létrehozásának ideje</th>
-                                <th>Igénylő személy</th>
-                                <th>Kinek a számára</th>
-                                <th>Igénylés típusa</th>
-                                <th>Státusz</th>
-                                <th></th>
-                            </tr>
-                        </thead>
-                        <tbody>
-                            @foreach (var item in Model)
-                            {
-                                <tr>
-                                    <td>
-                                        ISZR <span class="text-info font-weight-bold">#@Html.DisplayFor(modelItem => item.RequestId)</span>
-                                    </td>
-                                    <td>
-                                        @Html.DisplayFor(modelItem => item.CreationDate)
-                                    </td>
-                                    <td>
-                                        @{
-                                            if (item.RequestAuthor == null)
-                                            {
-                                                <span class="text-danger font-weight-bold">Törölt felhasználó</span>
-                                            }
-                                            else
-                                            {
-                                                <span>@item.RequestAuthor.DisplayName</span>
-                                            }
-                                        }
-                                    </td>
-                                    <td>
-                                        @{
-                                            if (item.RequestFor == null)
-                                            {
-                                                <span class="text-danger font-weight-bold">Törölt felhasználó</span>
-                                            }
-                                            else
-                                            {
-                                                <span>@item.RequestFor.DisplayName</span>
-                                            }
-                                        }
-                                    </td>
-                                    <td>
-                                        @Html.DisplayFor(modelItem => item.Type)
-                                    </td>
-                                    <td>
-                                        @{
-                                            if (item.Status == "Végrehajtva")
-                                            {
-                                                <span class="text-success font-weight-bold"><i class="fas fa-check"></i> Végrehajtva</span>
-                                            }
-                                            else if (item.Status == "Elutasítva")
-                                            {
-                                                <span class="text-danger font-weight-bold"><i class="fas fa-ban"></i> Elutasítva</span>
-                                            }
-                                            else
-                                            {
-                                                <span class="text-info font-weight-bold"><i class="fas fa-comment-dots"></i> Folyamatban</span>
-                                            }
-
-                                        }
-                                    </td>
-                                    <td class="float-right">
-                                        <a asp-action="Details" asp-route-id="@item.RequestId"><input type="button" value="Részletek" class="btn btn-info" /></a>
-                                    </td>
-                                </tr>
-                            }
-                        </tbody>
-                    </table>
-                </div>
-
-            </div>
-        </div>
-    </div>
-</div>
