+++ Views/Requests/RecordsByTime.cshtml
@@ -1,5 +1,150 @@
-﻿@*
-    For more information on enabling MVC for empty projects, visit https://go.microsoft.com/fwlink/?LinkID=397860
-*@
+﻿@model ISZR.Models.Request
+
 @{
+	ViewData["Title"] = "Kamerafelvétel lementése időpontok alapján";
 }
+
+<div class="container-fluid">
+	<div class="row">
+		<div class="col-md-12">
+			<div class="card card-default">
+				<div class="card-header bg-info">
+					<h3 class="card-title">Igénylés menete</h3>
+				</div>
+				<div class="card-body p-0">
+					<form asp-action="RecordsByTags">
+						<div class="row">
+							<div class="col-12">
+								<div asp-validation-summary="ModelOnly" class="text-danger"></div>
+							</div>
+						</div>
+						<div class="bs-stepper linear">
+							<div class="bs-stepper-header" role="tablist">
+								<div class="step active" data-target="#request-info">
+									<button type="button" class="step-trigger" role="tab" aria-controls="request-info" id="request-info-trigger" aria-selected="true">
+										<span class="bs-stepper-circle bg-info"><i class="fas fa-camera"></i></span>
+										<span class="bs-stepper-label">Felvételel kapcsolatos információk</span>
+									</button>
+								</div>
+								<div class="line"></div>
+								<div class="step" data-target="#recordTime">
+									<button type="button" class="step-trigger" role="tab" aria-controls="recordTime" id="recordTime-trigger" aria-selected="false" disabled="disabled">
+										<span class="bs-stepper-circle bg-info"><i class="fas fa-clock"></i></span>
+										<span class="bs-stepper-label">Kamerák és azok időpontjai</span>
+									</button>
+								</div>
+								<div class="line"></div>
+								<div class="step" data-target="#summary">
+									<button type="button" class="step-trigger" role="tab" aria-controls="summary" id="summary-trigger" aria-selected="false" disabled="disabled">
+										<span class="bs-stepper-circle bg-info"><i class="fas fa-list"></i></span>
+										<span class="bs-stepper-label">Összesítés</span>
+									</button>
+								</div>
+							</div>
+							<div class="bs-stepper-content">
+								<div id="request-info" class="content active dstepper-block" role="tabpanel" aria-labelledby="request-info-trigger">
+									<div class="row">
+										<div class="col-12">
+											<div class="alert alert-info alert-dismissible">
+												<h5><i class="icon fas fa-info"></i> Figyelem!</h5>
+												A bástyafalon elhelyezett kamerák amelyek külterületre néznek, 3 napig rögzítenek felvételeket. Minden más kamera amely nem külterületre néz, körülbelül 1 hónapig rögzít.
+											</div>
+										</div>
+									</div>
+									<div class="row">
+										<div class="col-12 col-md-6">
+											<div class="form-group">
+												<label>Esemény dátuma</label>
+												<div class="input-group mb-3">
+													<div class="input-group-prepend">
+														<span class="input-group-text"><i class="fas fa-calendar"></i></span>
+													</div>
+													<input id="inputDate" name="inputDate" type="date" min="@DateTime.Now.AddMonths(-1).AddDays(-14).ToString("yyyy-MM-dd")" max="@DateTime.Now.ToString("yyyy-MM-dd")" class="form-control" required="required" />
+												</div>
+											</div>
+										</div>
+										<div class="col-12 col-md-6">
+											<div class="form-group">
+												<label>Kamerafelvétel lementésének oka</label>
+												<div class="input-group mb-3">
+													<div class="input-group-prepend">
+														<span class="input-group-text"><i class="far fa-eye"></i></span>
+													</div>
+													<input id="inputWhy" name="inputWhy" type="text" class="form-control" required="required" />
+												</div>
+											</div>
+										</div>
+									</div>
+									<div class="row">
+										<div class="col-12">
+											<a class="btn btn-info" onclick="stepper.next()">Tovább</a>
+										</div>
+									</div>
+								</div>
+								<div id="recordTime" class="content" role="tabpanel" aria-labelledby="recordTime-trigger">
+									<div class="row">
+										<div class="col-12">
+											<a class="btn btn-warning" onclick="stepper.previous()">Elöző</a>
+											<a class="btn btn-info" onclick="summary()">Tovább</a>
+											<script>
+												function summary() {
+													document.getElementById("summaryDate").innerHTML = $('#inputDate').val();
+													document.getElementById("summaryWhy").innerHTML = $('#inputWhy').val();
+													document.getElementById("summaryTags").innerHTML = $('#inputTags').val();
+													document.getElementById("summaryCameras").innerHTML = $('#selectedCameras option:selected').toArray().map(item => item.text).join(", ");
+													stepper.next();
+												}
+											</script>
+										</div>
+									</div>
+								</div>
+								<div id="summary" class="content" role="tabpanel" aria-labelledby="summary-trigger">
+									<!-- TODO: User can define how much records want -->
+									<div class="row">
+										<div class="col-12">
+											<div class="alert alert-light alert-dismissible">
+												<h5><i class="icon fas fa-exclamation-triangle"></i> Összefoglaló áttekintése az igénylés létrehozása előtt</h5>
+												<dl class="row">
+													<dt class="col-sm-4"><i class="fas fa-tag"></i> Igénylés típusa</dt>
+													<dd class="col-sm-8">Kamerafelvétel lementése</dd>
+													<dt class="col-sm-4"><i class="fas fa-history"></i> Igénylés létrehozásának ideje</dt>
+													<dd class="col-sm-8">@DateTime.Now.ToString("yyyy. MMMM dd. HH:mm")</dd>
+													<dt class="col-sm-4"><i class="fas fa-calendar"></i> Esemény dátuma</dt>
+													<dd id="summaryDate" class="col-sm-8"></dd>
+													<dt class="col-sm-4"><i class="far fa-eye"></i> Lementésének oka</dt>
+													<dd id="summaryWhy" class="col-sm-8"></dd>
+													<dt class="col-sm-4"><i class="fas fa-tags"></i> Címkék megnevezése</dt>
+													<dd id="summaryTags" class="col-sm-8"></dd>
+													<dt class="col-sm-4"><i class="fas fa-video"></i> Megcímkézett kamerák</dt>
+													<dd id="summaryCameras" class="col-sm-8"></dd>
+												</dl>
+											</div>
+										</div>
+									</div>
+									<div class="custom-control custom-checkbox mb-3">
+										<input type="checkbox" class="custom-control-input" id="acceptRequest" required>
+										<label class="custom-control-label" for="acceptRequest">Az összefoglalót elolvastam és minden információ helyesen lett megadva.</label>
+										<div class="invalid-feedback">Az adatok helyességének ellenőrzése nélkül nem készülhet el az igénylés.</div>
+									</div>
+									<div class="row">
+										<div class="col-12">
+											<a class="btn btn-warning" onclick="stepper.previous()">Elöző</a>
+											<button type="submit" class="btn btn-info">Igénylés létrehozása</button>
+										</div>
+									</div>
+								</div>
+							</div>
+						</div>
+					</form>
+				</div>
+			</div>
+
+		</div>
+	</div>
+</div>
+
+<script type="text/javascript">
+	document.addEventListener('DOMContentLoaded', function () {
+		window.stepper = new Stepper(document.querySelector('.bs-stepper'));
+	});
+</script>
