+++ Views/Shared/_Navbar.cshtml
@@ -25,9 +25,9 @@
 		@if (Account.IsAdmin())
 		{
 			<li class="nav-item ml-2">
-				<a class="nav-link btn btn-danger text-white" asp-controller="Administrator" asp-action="Index">
-					<i class="fas fa-pencil-alt"></i>
-					<span>Adminisztráció</span>
+				<a class="nav-link btn btn-danger text-white" asp-controller="Users" asp-action="Index">
+					<i class="fas fa-toolbox"></i>
+					<span>Karbantartás</span>
 				</a>
 			</li>
 		}
